Fix null reference in options panel
If the Preferences dialog is opened and the file watcher filter
options panel is not displayed then the UI controls are not
created. Then when the options panel is disposed it was
throwing a null reference trying to remove event handlers
from the UI controls. Added a null check to the Dispose method.

Fixes #1