Fix null reference on closing options panel dialog
It does not seem possible to remove the Xwt.ListView's
SelectionChanged event handler. When the options panel
Dispose method is called the underlying Gtk.TreeView's
Selection (TreeSelection type) is null. XWT throws a null
reference exception here since the event handler is on the
TreeSelection. Avoid this by not removing the event handler.